# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Social-Media-Api
  version: v1
  description: "Спецификация для проекта Social-Media-Api."
servers:
  - url: http://localhost:8080
    description: local server
paths:
...
  /api/v1/auth/login:
    post:
      tags:
        - Authentification endpoints
      summary: "Аутентификация пользователя"
      operationId: login
      requestBody:
        description: "Ввод имени и пароля пользователя для прохождения аутентификации"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JwtRequest'
        required: true
      responses:
        '200':
          description: Successful Auth
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/register:
    post:
      tags:
        - Authentification endpoints
      summary: "Регистрация пользователя"
      operationId: register
      requestBody:
        description: "Ввод имени, пароля и электронной почты для регистрации пользователя"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        '200':
          description: Successful registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentification endpoints
      summary: "Обновление токена аутентификации и токена обновления"
      operationId: refresh
      requestBody:
        description: "Ввод токена обновления"
        content:
          application/json:
            schema:
              type: string
              example: "$2a$10$j1PJv6GskindBDL2ZtyOj.gCzrMVx57/HeKp0zOBdd1.F1lWoseU6"
        required: true
      responses:
        '200':
          description: Successful refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JwtRequest:
      required:
        - name
        - password
      type: object
      properties:
        name:
          type: string
          description: "Имя пользователя"
          example: "Sergey"
        password:
          type: string
          description: "Пароль пользователя"
          example: "111"
    JwtResponse:
      required:
        - id
        - name
        - accessToken
        - refreshToken
      type: object
      properties:
        id:
          type: integer
          description: "Id пользователя"
          example: "1"
        name:
          type: string
          description: "Имя пользователя"
          example: "Sergey"
        accessToken:
          type: string
          description: "Токен аутентификации"
          example: "$2a$12$uhKy.MGqmfYcgbkTd3ZR.eyXKVrU3gm0epHhSy6cfXzydbX4n3Ws2"
        refreshToken:
          type: string
          description: "Токен обновления токена аутентификации"
          example: "$2a$12$5lWa/FKZ925t3p.QDvj/oeo38ot7Lna9vhhiCYCOBv8YV0h2uN/5q"
    UserDto:
      required:
        - id
        - email
        - name
        - password
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: "Id пользователя"
          example: "1"
        name:
          type: string
          description: "Имя пользователя"
          example: "Sergey"
        password:
          type: string
          description: "Пароль пользователя"
          example: "111"
          writeOnly: true
        email:
          type: string
          description: "Электронная почта пользователя"
          example: "Sergey@gmail.com"
    ErrorResponse:
      type: object
      properties:
        messageError:
          type: string
          description: Error message
          example: "Error"